Use monorepo model Rename to "waifu-picker"
