+++ package.json
@@ -1,9 +1,19 @@
 {
-  "name": "request-response",
+  "name": "waifu-picker",
   "packageManager": "yarn@4.5.1",
+  "workspaces": [
+    "src/*"
+  ],
+  "scripts": {
+    "build": "webpack",
+    "watch": "webpack --watch",
+    "start": "nodemon ./dist/server.js"
+  },
+  "dependencies": {
+    "reflect-metadata": "^0.2.2"
+  },
   "devDependencies": {
     "@eslint/js": "^9.12.0",
-    "@types/express": "^5.0.0",
     "@types/node": "^22.7.5",
     "@types/webpack-node-externals": "^3",
     "css-loader": "^7.1.2",
@@ -28,15 +38,5 @@
     "webpack-dev-server": "^5.1.0",
     "webpack-merge": "^6.0.1",
     "webpack-node-externals": "^3.0.0"
-  },
-  "scripts": {
-    "build": "webpack",
-    "watch": "webpack --watch",
-    "start": "nodemon ./dist/server.js"
-  },
-  "dependencies": {
-    "express": "^4.21.1",
-    "reflect-metadata": "^0.2.2",
-    "tsyringe": "^4.8.0"
   }
 }
