+++ src/client/components/app-navigation/navigation.scss
@@ -1,4 +1,5 @@
-@use "common.scss";
+@use "styles/variables.scss" as vars;
+@use "styles/mixins.scss" as mx;
 
 .app-nav {
   display: flex;
@@ -7,17 +8,14 @@
   margin-bottom: 0.5em;
   border-top: solid darkorange 3px;
   border-bottom: solid darkorange 3px;
-  @media screen and (max-width: common.$vertical-width) {
+
+  @media screen and (max-width: vars.$vertical-width) {
     justify-content: space-evenly;
   }
 
   .nav-item {
-    @include common.button-interact();
-    @media screen and (min-width: common.$vertical-width) {
-      margin-inline: 2em;
-    }
+    @include mx.button-interact();
 
-    filter: drop-shadow(1px 1px 1px grey);
     min-width: max(150px, 15%);
     padding: 0.5em;
     border-radius: 0.5em;
@@ -27,6 +25,11 @@
     text-align: center;
     background-color: darkmagenta;
     color: white;
+    filter: drop-shadow(1px 1px 1px grey);
+
+    @media screen and (min-width: vars.$vertical-width) {
+      margin-inline: 2em;
+    }
   }
 }
