+++ src/client/components/page-chars/character-card/character.card.scss
@@ -1,4 +1,4 @@
-@use "common.scss";
+@use "styles/variables.scss" as vars;
 
 .character {
   display: flex;
@@ -10,7 +10,7 @@
   width: fit-content;
   max-width: 640px;
   min-height: 520px;
-  @media screen and (min-width: common.$vertical-width) {
+  @media screen and (min-width: vars.$vertical-width) {
     margin-left: 5em;
   }
 
@@ -22,7 +22,7 @@
     border-radius: 2em;
     padding: 1em;
 
-    @media screen and (max-width: common.$vertical-width) {
+    @media screen and (max-width: vars.$vertical-width) {
       flex-direction: column;
     }
 
@@ -37,7 +37,7 @@
       filter: drop-shadow(5px 5px 4px dimgrey);
       border-radius: 2em;
       background-color: white;
-      @media screen and (max-width: common.$vertical-width) {
+      @media screen and (max-width: vars.$vertical-width) {
         align-self: center;
         width: 400px;
         object-fit: scale-down;
@@ -50,7 +50,7 @@
       flex-direction: column;
       justify-content: space-evenly;
       margin-left: 1em;
-      @media screen and (max-width: common.$vertical-width) {
+      @media screen and (max-width: vars.$vertical-width) {
         margin-left: 0;
       }
 
@@ -113,7 +113,7 @@
         width: 100%;
         min-width: 300px;
         margin-top: 1em;
-        @media screen and (max-width: common.$vertical-width) {
+        @media screen and (max-width: vars.$vertical-width) {
           align-self: center;
           margin-top: 1em;
         }
