+++ src/client/components/page-chars/characters/character.list.html
