+++ src/client/components/page-chars/visionColors.ts
@@ -1,0 +1,13 @@
+/**
+ * Known values representing colors for character's elements.
+ */
+export enum VisionColors {
+  pyro = "red",
+  hydro = "blue",
+  dendro = "lightgreen",
+  geo = "orange",
+  cryo = "lightblue",
+  anemo = "teal",
+  electro = "blueviolet",
+  default = "golden",
+}
