+++ src/client/components/page-pols/polling/polling.ts
@@ -1,5 +1,6 @@
 import "./polling.scss";
 import template from "./polling.html";
+
 import { injectable } from "tsyringe";
 import { ComponentBase } from "@client/components/components";
