+++ src/client/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@wp/client",
+  "version": "0.4.0",
+  "packageManager": "yarn@4.5.1",
+  "type": "module",
+  "dependencies": {
+    "@wp/common": "workspace:^",
+    "reflect-metadata": "^0.2.2",
+    "tsyringe": "^4.8.0"
+  }
+}
