+++ src/client/styles/mixins.scss
@@ -1,0 +1,13 @@
+@mixin button-interact {
+  border: 0;
+
+  &:hover {
+    filter: brightness(115%);
+  }
+  &:active {
+    filter: brightness(95%);
+  }
+  &:focus {
+    outline: none;
+  }
+}
