+++ src/client/styles/variables.scss
@@ -1,0 +1,2 @@
+$vertical-width: 640px;
+$split-width: 1280px;
