+++ src/common/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@wp/common",
+  "version": "0.2.0",
+  "packageManager": "yarn@4.5.1",
+  "type": "module"
+}
